:recycle:(chore) refactor routes logic to controllers
