+++ server/src/modules/appointments/infra/http/v1/routes/appointments/index.ts
@@ -1,0 +1,14 @@
+import { Router } from 'express';
+
+import ensureAuthentication from '~/modules/users/infra/http/v1/middlewares/ensureAuthentication';
+
+import AppointmentsController from '../../controllers/AppointmentsController';
+
+const routes = Router();
+routes.use(ensureAuthentication);
+
+const appointmentsController = new AppointmentsController();
+
+routes.post('/', appointmentsController.create);
+
+export default routes;
