+++ server/src/modules/users/infra/http/v1/middlewares/ensureAuthentication.ts
@@ -1,8 +1,8 @@
 import { StatusCodes } from 'http-status-codes';
 import { verify } from 'jsonwebtoken';
 
-import authConfig from '@config/auth-config';
-import AppException from '@shared/exceptions/AppException';
+import authConfig from '~/config/auth-config';
+import AppException from '~/shared/exceptions/AppException';
 import { ExpressMiddleware } from '~/@types/middleware';
 
 interface TokenPayload {
