+++ server/src/modules/users/infra/http/v1/routes/sessions/index.ts
@@ -1,0 +1,11 @@
+import { Router } from 'express';
+
+import SessionsController from '../../controllers/SessionsController';
+
+const routes = Router();
+
+const sessionsController = new SessionsController();
+
+routes.post('/', sessionsController.create);
+
+export default routes;
