+++ server/src/shared/infra/http/routes/v1/index.ts
@@ -1,8 +1,8 @@
 import { Router } from 'express';
 
-import appointmentsRouter from '@modules/appointments/infra/http/routes/v1/appointments';
-import usersRouter from '@modules/users/infra/http/routes/v1/users';
-import sessionsRouter from '@modules/users/infra/http/routes/v1/sessions';
+import appointmentsRouter from '~/modules/appointments/infra/http/v1/routes/appointments';
+import usersRouter from '~/modules/users/infra/http/v1/routes/users';
+import sessionsRouter from '~/modules/users/infra/http/v1/routes/sessions';
 
 const routes = Router();
